# unidad4/swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "Concesionarios API.",
    "version": "1.0.2"
  },
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "marca": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        }
      },
      "required": ["marca", "modelo"]
    },
    "Concesionarios": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "coches": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Coche"
          }
        }
      },
      "required": ["nombre", "direccion"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Fetch concesionarios",
        "operationId": "getConcesionarios",
        "responses": {
          "200": {
            "description": "List of concesionarios.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionarios"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create concesionarios",
        "operationId": "createConcesionarios",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "concesionarios",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Get a concesionarios by id.",
        "operationId": "getConcesionarios",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch concesionarios.",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        }
      },
      "put": {
        "summary": "Update concesionarios",
        "operationId": "updateConcesionarios",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "concesionarios",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      },
      "delete": {
        "summary": "Delete concesionarios.",
        "operationId": "deleteConcesionarios",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Fetch coches",
        "operationId": "getCoches",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of coches.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Concesionarios"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create coches",
        "operationId": "createCoches",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "$ref": "#/definitions/Coche"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      }
    },
    "/concesionarios/{id}/coches/{id}": {
      "get": {
        "summary": "Get a coches by id.",
        "operationId": "getCoches",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch coches.",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        }
      },
      "put": {
        "summary": "Update coches",
        "operationId": "updateCoches",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "coches",
            "schema": {
              "$ref": "#/definitions/Concesionarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      },
      "delete": {
        "summary": "Delete coches.",
        "operationId": "deleteCoches",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "{ message: 'ok' }"
          }
        }
      }
    }
  }
}